firmware: arm_scmi: add regulator support based on SCMI
Add a simple regulator based on SCMI Voltage Domain Protocol.

A restricted set of the usual regulator operations is currently supported,
since one of the stated aims of the chosen backend protocol is to limit
the amount of operational details available and configurable from Kernel
space.

Supported operations are:
- .enable / .disable / .is_enabled: routed via SCMI Voltage Domain Protocol
- .get_voltage / .set_voltage: routed via SCMI Voltage Domain Protocol
- .get_voltage_sel/.set_voltage_sel: using regulator framework helpers
- .list_voltage/.map_voltage: using regulator frameowrk helpers

Given the limited scope of the operations supported and the abstract nature
of the backend protocol, only a minimal subset of the standard regulator
configuration bindings are effectively considered, while most of them are
ignored.

Signed-off-by: Cristian Marussi <cristian.marussi@arm.com>