scmi/optee: support multiple message per channel
In order to support more that 1 request at the same time per channel,
we must increase the size of the allocated shared memory and split it
in several different message. We map the pre-allocated scmi message id
to the offset in the shared memory.

The scmi pre-allocated id is saved in xfer->hdr.seq

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
[hacked etienne: fixup coding style, build warnings, fix process_event()]
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>